+++ packages/categorize/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.12](https://github.com/pie-framework/pie-ui/compare/@pie-ui/categorize@0.9.11...@pie-ui/categorize@0.9.12) (2019-04-25)
+
+**Note:** Version bump only for package @pie-ui/categorize
+
+
+
+
+
+## [0.9.11](https://github.com/pie-framework/pie-ui/compare/@pie-ui/categorize@0.9.10...@pie-ui/categorize@0.9.11) (2019-04-25)
+
+**Note:** Version bump only for package @pie-ui/categorize
+
+
+
+
+
+## [0.9.10](https://github.com/pie-framework/pie-ui/compare/@pie-ui/categorize@0.9.9...@pie-ui/categorize@0.9.10) (2019-04-15)
+
+**Note:** Version bump only for package @pie-ui/categorize
+
+
+
+
+
+## [0.9.9](https://github.com/pie-framework/pie-ui/compare/@pie-ui/categorize@0.9.8...@pie-ui/categorize@0.9.9) (2019-04-10)
+
+**Note:** Version bump only for package @pie-ui/categorize
+
+
+
+
+
 ## [0.9.8](https://github.com/pie-framework/pie-ui/compare/@pie-ui/categorize@0.9.7...@pie-ui/categorize@0.9.8) (2019-04-09)
 
 **Note:** Version bump only for package @pie-ui/categorize