+++ packages/text-entry/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.5.9](https://github.com/pie-framework/pie-ui/compare/@pie-ui/text-entry@3.5.8...@pie-ui/text-entry@3.5.9) (2019-04-25)
+
+**Note:** Version bump only for package @pie-ui/text-entry
+
+
+
+
+
+## [3.5.8](https://github.com/pie-framework/pie-ui/compare/@pie-ui/text-entry@3.5.7...@pie-ui/text-entry@3.5.8) (2019-04-25)
+
+**Note:** Version bump only for package @pie-ui/text-entry
+
+
+
+
+
+## [3.5.7](https://github.com/pie-framework/pie-ui/compare/@pie-ui/text-entry@3.5.6...@pie-ui/text-entry@3.5.7) (2019-04-18)
+
+**Note:** Version bump only for package @pie-ui/text-entry
+
+
+
+
+
 ## [3.5.6](https://github.com/pie-framework/pie-ui/compare/@pie-ui/text-entry@3.5.5...@pie-ui/text-entry@3.5.6) (2019-04-04)
 
 **Note:** Version bump only for package @pie-ui/text-entry